HandsOn 2
In this exercise, Helm was used to package and deploy an Nginx application in Kubernetes. A Helm chart was created, and application configuration such as replica count and image version was defined in the values.yaml file. The chart was installed as a Helm release, creating Kubernetes resources automatically. The application was upgraded using helm upgrade, demonstrating rolling updates, and then rolled back to a previous version using helm rollback. Finally, the application was completely removed using helm uninstall, showcasing Helm’s lifecycle management capabilities.